requirements.txt
# Flask and web framework
flask
flask-sqlalchemy
flask-cors

# Computer vision and image processing
opencv-python==4.10.0.84
numpy==1.26.4
pillow
scikit-learn
pytesseract

# Task queue and caching
celery
redis

# Database and environment
psycopg2-binary
python-dotenv

# Testing
pytest
